Remove a warning for conversion double->long.  This has impacts on Windows. PR: 849

+++ crypto/bio/b_print.c
@@ -641,9 +641,9 @@ fmtfp(
        multiplying by a factor of 10 */
     fracpart = roundv((pow10(max)) * (ufvalue - intpart));
 
-    if (fracpart >= pow10(max)) {
+    if (fracpart >= (long)pow10(max)) {
         intpart++;
-        fracpart -= pow10(max);
+        fracpart -= (long)pow10(max);
     }
 
     /* convert integer part */
